+++ nutella/src/container/Mypage/index.tsx
@@ -1,6 +1,6 @@
 import Footer from "../../components/Footer/Footer";
 import Header from "../../components/Header/Header";
-import MyPage from "../../components/Mypage/Mypage";
+import MyPage from "../../components/Profile/Mypage/Mypage";
 
 const MypageContainer = () => {
   return (
